Changing subview container description
The current docs state:
> container {String} Selector that describes the element within the view that should hold the subview
To me, that indicates that the container will remain, but have a new child once the subview is rendered.  But this is not happening.  Instead, ampersand-view.renderWithTemplate() is replacing the container with the subview:
```javascript
var parent = this.el && this.el.parentNode;
if (parent) parent.replaceChild(newDom, this.el);
```
This changes the wording about how the container is used.